isolating the findCommandByKeys function into a single module so tests stop crashing due to importing ../config
also created more tests in order to test each module individually
