+++ app/utils/keymaps-find-command-by-keys.js
@@ -1,0 +1,5 @@
+const normalize = require('./keymaps-normalize');
+
+module.exports = (keys, commands) => {
+  return commands[normalize(keys)];
+};
