+++ lib/utils/keymaps.js
@@ -1,6 +1,6 @@
 import {remote} from 'electron';
 
-const {getCommand} = remote.require('./utils/keymaps-normalize');
+const {getCommand} = remote.require('./utils/keymaps-get-command');
 
 export default function returnKey(e) {
   let keys = [];
