+++ test/unit/keymaps-normalize.test.js
@@ -1,64 +1,22 @@
 import test from 'ava';
-import {findCommandByKeys} from '../../app/utils/keymaps-normalize';
+import normalize from '../../app/utils/keymaps-normalize';
 
-const expectedCommand = 'test-command';
-const expectedLocalizedCommand = 'test-localized-command';
-
-const commands = {
-  'alt+p+shift': expectedCommand,
-  'ç+cmd+ctrl': expectedLocalizedCommand
-};
-
-test(`returns a command`, t => {
-  t.is(
-    findCommandByKeys('alt+shift+p', commands),
-    expectedCommand
-  );
+test(`is normalizing keymaps correctly`, t => {
+  const notNormalized = 'cmd+alt+o';
+  const normalized = 'alt+cmd+o';
 
   t.is(
-    findCommandByKeys('shift+p+alt', commands),
-    expectedCommand
-  );
-
-  t.is(
-    findCommandByKeys('p+alt+shift', commands),
-    expectedCommand
-  );
-
-  t.is(
-    findCommandByKeys('alt+shift+P', commands),
-    expectedCommand
-  );
-
-  t.is(
-    findCommandByKeys('Shift+P+Alt', commands),
-    expectedCommand
+    normalize(notNormalized),
+    normalized
   );
 });
 
-test(`returns a localized command`, t => {
-  t.is(
-    findCommandByKeys('cmd+ctrl+ç', commands),
-    expectedLocalizedCommand
-  );
+test(`is normalizing localized keymaps correctly`, t => {
+  const notNormalized = 'cmd+alt+ç';
+  const normalized = 'alt+ç+cmd';
 
   t.is(
-    findCommandByKeys('ç+cmd+ctrl', commands),
-    expectedLocalizedCommand
-  );
-
-  t.is(
-    findCommandByKeys('ctrl+ç+cmd', commands),
-    expectedLocalizedCommand
-  );
-
-  t.is(
-    findCommandByKeys('ctrl+Ç+cmd', commands),
-    expectedLocalizedCommand
-  );
-
-  t.is(
-    findCommandByKeys('Cmd+Ctrl+Ç', commands),
-    expectedLocalizedCommand
+    normalize(notNormalized),
+    normalized
   );
 });
